pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "setuptools-scm>=8.0"]
build-backend = "setuptools.build_meta"

[project]
name = "crazypumpkin"
version = "0.1.0"
description = "Autonomous AI Company OS — install, configure, and run your own AI-powered software company"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    {name = "Majd Kassawat"},
]
keywords = ["ai", "agents", "autonomous", "company", "framework", "multi-agent"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Application Frameworks",
    "Operating System :: OS Independent",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "pyyaml>=6.0",
    "fastapi>=0.100",
    "uvicorn>=0.20",
    "httpx>=0.25",
    "anthropic>=0.30",
    "jinja2>=3.1",
    "watchdog>=3.0",
]

[project.optional-dependencies]
openai = ["openai>=1.0"]
voice = ["openai>=1.0"]
telegram = ["python-multipart>=0.0.6"]
all = ["openai>=1.0", "python-multipart>=0.0.6"]
plugins = ["importlib-metadata>=6.0; python_version<'3.12'"]
tracing = ["langfuse>=2.0"]
dev = ["pytest>=7.0", "pytest-asyncio>=0.21"]

# Third-party packages register CP-OS plugins via entry-points.
# Format: plugin_name = "module.path:PluginClass"
[project.entry-points."crazypumpkin.plugins"]

[project.scripts]
crazy-pumpkin = "crazypumpkin.cli:main"

[project.urls]
Homepage = "https://github.com/majdkassawat/crazy-pumpkin-os"
Documentation = "https://github.com/majdkassawat/crazy-pumpkin-os/tree/main/docs"
Repository = "https://github.com/majdkassawat/crazy-pumpkin-os"
Issues = "https://github.com/majdkassawat/crazy-pumpkin-os/issues"
Changelog = "https://github.com/majdkassawat/crazy-pumpkin-os/releases"

[tool.pytest.ini_options]
pythonpath = ["src"]

[tool.setuptools.packages.find]
where = ["src"]
